Add colours to the game
We just have a lookup table to map TetriType to sf::Colour and a helper
function to get that value.

The GameGrid must now store an optional TetriType so that we can
have multiple colours.
We store an std::optional<Tetritype> to make it easier to test for their
presence.